+++ UVtools.WPF/Windows/AboutWindow.axaml.cs
@@ -20,9 +20,9 @@ namespace UVtools.WPF.Windows
         public string FrameworkDescription => RuntimeInformation.FrameworkDescription;
         public int ProcessorCount => Environment.ProcessorCount;
         public int ScreenCount => Screens.ScreenCount;
-        public string ScreenResolution => $"{Screens.Primary.Bounds.Width} x {Screens.Primary.Bounds.Height} @ {Screens.Primary.PixelDensity*100}%";
-        public string WorkingArea => $"{Screens.Primary.WorkingArea.Width} x {Screens.Primary.WorkingArea.Height}";
-        public string RealWorkingArea => $"{App.MaxWindowSize.Width} x {App.MaxWindowSize.Height}";
+        //public string ScreenResolution => $"{Screens.Primary.Bounds.Width} x {Screens.Primary.Bounds.Height} @ {Screens.Primary.PixelDensity*100}%";
+        //public string WorkingArea => $"{Screens.Primary.WorkingArea.Width} x {Screens.Primary.WorkingArea.Height}";
+        //public string RealWorkingArea => $"{App.MaxWindowSize.Width} x {App.MaxWindowSize.Height}";
 
         public string ScreensDescription
         {
@@ -31,8 +31,12 @@ namespace UVtools.WPF.Windows
                 var result = new StringBuilder();
                 for (int i = 0; i < Screens.All.Count; i++)
                 {
+                    var onScreen = Screens.ScreenFromVisual(App.MainWindow);
                     var screen = Screens.All[i];
-                    result.AppendLine($"{i+1}: {screen.Bounds.Width} x {screen.Bounds.Height} @ {screen.PixelDensity * 100}%");
+                    result.AppendLine($"{i+1}: {screen.Bounds.Width} x {screen.Bounds.Height} @ {screen.PixelDensity * 100}%" + 
+                        (screen.Primary ? " (Primary)" : string.Empty) +
+                        (onScreen == screen ? " (On this)" : string.Empty)                        
+                        );
                     result.AppendLine($"    WA: {screen.WorkingArea.Width} x {screen.WorkingArea.Height}    UA: {Math.Round(screen.WorkingArea.Width / screen.PixelDensity)} x {Math.Round(screen.WorkingArea.Height / screen.PixelDensity)}");
                 }
                 return result.ToString().TrimEnd();